+++ test/flow/window_test.exs
@@ -6,18 +6,18 @@ defmodule Flow.Window.Test do
 
   test "periodic triggers" do
     assert Flow.Window.global
-           |> Flow.Window.trigger_periodically(10, :seconds, :keep)
+           |> Flow.Window.trigger_periodically(10, :second, :keep)
            |> Map.fetch!(:periodically) ==
-           [{10000, :keep, {:periodically, 10, :seconds}}]
+           [{10000, :keep, {:periodically, 10, :second}}]
 
     assert Flow.Window.global
-           |> Flow.Window.trigger_periodically(10, :minutes, :keep)
+           |> Flow.Window.trigger_periodically(10, :minute, :keep)
            |> Map.fetch!(:periodically) ==
-           [{600000, :keep, {:periodically, 10, :minutes}}]
+           [{600000, :keep, {:periodically, 10, :minute}}]
 
     assert Flow.Window.global
-           |> Flow.Window.trigger_periodically(10, :hours, :keep)
+           |> Flow.Window.trigger_periodically(10, :hour, :keep)
            |> Map.fetch!(:periodically) ==
-           [{36000000, :keep, {:periodically, 10, :hours}}]
+           [{36000000, :keep, {:periodically, 10, :hour}}]
   end
 end